fix: middlewares refactor
